[chore] Cra to Vite migration
