--- film_journey/.gitignore
@@ -1,26 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-firebase.js
-
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# production
-/build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
