--- film_journey/src/index.js
@@ -1,20 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import { ThemeProvider } from "@mui/material/styles";
-import theme from "../src/styles/theme";
-import { CssBaseline } from "@mui/material";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-	<ThemeProvider theme={theme}>
-		<CssBaseline /> {/* 기본 CSS 초기화 */}
-		<App />
-	</ThemeProvider>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
